blockbook: add project skelleton
Since this is a public library project we want to pin
down the smallest required dependency versions. Also,
we use `thiserror` instead of `anyhow`.

Even though all endpoints start with `/api/v2/` we don't
consider them to be part of the base URL to be consistent
with the blockbook documentation.

Integration tests instead of unit tests are implemented
to better experience the API user's perspective ourselves.