blockbook: derive Clone and Serialize
It is obvious that a library user wants to be able
to clone the objects defined in this crate. Therefore,
we derive the `Clone` trait. Since we have a hard
dependency on `serde` we can as well also derive
`serde::Serialize` which is likely to be useful, too.

The `serde::Serialize` derive required to implement
a `serialize()` function for `Amount` which we now
unit test for both de- and serialization.