blockbook/test: timeout instances individually
Blockbook testing instances should observe a timeout
period before they get reused for another test.

The previous mechanism to ensure this relied on a brittle
counting mechanism, and blocked the entire collection to
wait for the longest-running instance, before making the
entire collection available again.

This commit replaces this with a FIFO queue of instances
that indicate at which time point they become available again.

The nested scope when fetching a blockbook instance from the
front of the queue is necessary to ensure that the mutex guard
is dropped before any `await` points are encountered.

Storing the blockbook instance in a reference-counted field
allows it to be cheaply cloned in `drop()`. The referece count
needs to be atomic because of its use within `Lazy`.