blockbook: implement block() methods
These methods wrap the following blockbook endpoint, see docs:
https://github.com/trezor/blockbook/blob/86ff5a9538dba6b869f53850676f9edfc3cb5fa8/docs/api.md#get-block

The reason why u8 is not big enough for the version field in the
Block struct is that is is being repurposed for soft fork signaling.
See:
https://docs.rs/bitcoin/latest/bitcoin/blockdata/block/struct.BlockHeader.html#structfield.version
For an example see block 500044.

An example of why the Option<> around the spent field in the
BlockVout struct is needed is the genesis block.

The AddressBlockVout enum is needed as the bitcoin library does not
support serde deserialization of OP_RETURN variants.

Also, for this commit the blockbook timeout had to be adjusted for it
to be able to pass the ci pipeline.