21: ban use of Default::default()
We have doubted in clippy's wisdom but have been thought
otherwise. This is a hard lessons learned. Quoting clippy:

> It’s easier for the reader if the name of the type
> is used, rather than the generic Default.

On top of that we've noticed, that sometimes we've used
`Default::default()` instead of the more succinct `None`
which would have been noticed immediately with the lint
enabled.

We also eliminate a `Default` derive which was used in
exactly one test only. The test gained in verbosity but
also in clarity.