blockbook: implement address_info() methods
The serde(deny_unknown_fields) attribute had to be removed for some
structs. The reason for this is the mentioned structs are used as
fields by new structs introduced in this commit. The field structs
do not fit perfectly, meaning they have a field less than the
actually returned json data which additionally includes an optional
field called `isOwn`. In addition to that, the macro is not supported
in combination with the flatten macro.

Note: Any returned transaction iterable by the API is in reverse
chronological order meaning the newest transaction comes at the first
index in the iterable. Also, when using the `page` argument, lower
pages yield newer transactions again starting with the newest
transaction in the iterable.