blockbook: implement utxo() methods
Unfortunately the bitcoin library can't parse zpubs, so we're forced
to use `String` or `&str` as arguments instead of a proper type that
does additional parsing checks. We'll leave this until we find a
fitting library also supporting ypubs and zpubs.

For one of the two integration tests we burned a few satoshi to have
a static utxo and for the other integration test, to have some
diversity, we used the first utxo of the counterparty burn address
for which nobody has the corresponding private key, i.e. its utxos
will not be spent either.