blockbook/ws: handle exited processing loop gracefully
This unwrap turns out to be too eager: The processing
loop may have already exited due to encountered websocket
errors (L237) or a closed connection (L222).

In any case, the receiving end of the shutdown channel is
only dropped when the `process` method returns, which
means that the sender's intention was already accomplished.

This statemet also panicked for unrelated reasons in the
test suite, as can be observed by running tests with
`-- --nocapture`. There, it was presumably due to the fact
that the tokio runtime had already cancelled all pending tasks
before the const `QUEUED_BLOCKBOOKS_WS` were deallocated.