blockbook: implement xpub_info() methods
Note that the filter arguments `page`, `page_size`, `from` and `to`
only affect the `txids` and `transactions` arrays and leave the
`tokens` array alone.

To reuse code, the `Transaction` struct from the `transaction()`
method was also used for `xpub_info()`. Said struct however is
missing a field within `Vin` when exactly matched against the
returned data from `xpub_info()`. As a consequence the incoming
script in `Vin` is discarded. If the script turns out to be desired
information, it could be added as an optional field in `Transaction`
in a succeeding commit.

There are two reasons the fields `tokens`, `txids` and `transactions`
are required to be optional:

- The struct XpubInfoBasic is reused for both detail levels of
`include_token_list` respectively the struct XpubInfo is reused for
both detail levels of `entire_txs`.

- The other reason is independent of our datastructure design. Namely
in case of the fields being empty due to query specifications/
arguments, the blockbook API omits them entirely instead of returning
empty arrays.

Also note that the implementation of the `tokens` variant of the
`details` parameter was omitted for design simplicity.