blockbook: support queries on unused addresses
Unfortunately, inexistent transactions or transaction
ids are represented by an absent field, rather than an
empty list value.

As a consequence, it is not possible to dissociate the
`AddressInfo` and `AddressInfoDetailed` structs anymore
when deserializing without the usage of `deny_unknown_fields`,
which we do not want to activate for production builds.

The two structs are therefore merged into one, and it is
the caller's responsibility to pick out the return field
that contains the data he requested.

[1] https://blockbook.21analytics.xyz/api/v2/address/1LMx3bUxSKTxkLJxrYdku4Vv55JHifA12V