blockbook: catch version discrepancy in client initialization
This will alert the user to a potential version discrepancy thereby
eliminating the need for associated debugging.

Note: We ignore the version check in `test_status()` and
`test_info_ws()` as this is now already done in the initialization of
the Blockbook clients. Like so, the diff will stay smaller in a
potential future version bump while maintaining the same test
coverage.