Update rust version to v1.75.0
Rust 1.75.0 has stabilized async traits which
allows us to drop the `async-trait` dependency.

Also, `cargo` doesn't force us to specify a
`version` in the manifest anymore. Hence, we
have removed that field for unpublished crates
to imply that the `version` is unused. Because
the `version` field is required for publishing
a crate `public = false` has become redundant
and is removed as well.

Last, we have fixed the `get_first` clippy lint
and suppressed the `struct_field_names` lint.